+++ custom/requisitions/readme.md
@@ -1,0 +1,6 @@
+Legacy Requisitions App
+======================
+
+This app exists for legacy support for requisitions in CommTrack.
+Currently (as of July 2015) this is treated as custom code and a candidate for deletion.
+No core code should depend on anything in here.
